+++ packages/table/src/cell/formats/truncatedFormat2.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Palantir Technologies, Inc. All rights reserved.
+ * Copyright 2022 Palantir Technologies, Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,16 @@
  * limitations under the License.
  */
 
-/**
- * @fileoverview This component is DEPRECATED, and the code is frozen.
- * All changes & bugfixes should be made to TruncatedFormat2 instead.
- */
-
-/* eslint-disable deprecation/deprecation, @blueprintjs/no-deprecated-components */
-
 import classNames from "classnames";
 import * as React from "react";
 
-import { DISPLAYNAME_PREFIX, Icon, Popover, Position, Props } from "@blueprintjs/core";
+import { DISPLAYNAME_PREFIX, Icon } from "@blueprintjs/core";
+import { Popover2 } from "@blueprintjs/popover2";
 
 import * as Classes from "../../common/classes";
 import { Utils } from "../../common/utils";
 import { Locator } from "../../locator";
+import { ITruncatedFormatState, TruncatedFormatProps, TruncatedPopoverMode } from "./truncatedFormat";
 
 // amount in pixels that the content div width changes when truncated vs when
 // not truncated. Note: could be modified by styles
@@ -36,131 +31,8 @@ import { Locator } from "../../locator";
 // on the div, which comes from styles
 const CONTENT_DIV_WIDTH_DELTA = 25;
 
-export enum TruncatedPopoverMode {
-    ALWAYS = "always",
-    NEVER = "never",
-    WHEN_TRUNCATED = "when-truncated",
-    WHEN_TRUNCATED_APPROX = "when-truncated-approx",
-}
-
-export interface ITrucatedFormateMeasureByApproximateOptions {
-    /**
-     * Approximate character width (in pixels), used to determine whether to display the popover in approx truncation mode.
-     * The default value should work for normal table styles,
-     * but should be changed as necessary if the fonts or styles are changed significantly.
-     *
-     * @default 8
-     */
-    approximateCharWidth: number;
-
-    /**
-     * Approximate line height (in pixels), used to determine whether to display the popover in approx truncation mode.
-     * The default value should work for normal table styles, but should be changed if the fonts or styles are changed significantly.
-     *
-     * @default 18
-     */
-    approximateLineHeight: number;
-
-    /**
-     * Total horizonal cell padding (both sides), used to determine whether to display the popover in approx truncation mode.
-     * The default value should work for normal table styles,
-     * but should be changed as necessary if the fonts or styles are changed significantly.
-     *
-     * @default 20
-     */
-    cellHorizontalPadding: number;
-
-    /**
-     * Number of buffer lines desired, used to determine whether to display the popover in approx truncation mode.
-     * Buffer lines are extra lines at the bottom of the cell that space is made for, to make sure that the cell text will fit
-     * after the math calculates how many lines the text is expected to take.
-     * The default value should work for normal table styles,
-     * but should be changed as necessary if the fonts or styles are changed significantly.
-     *
-     * @default 0
-     */
-    numBufferLines: number;
-}
-
-export type TruncatedFormatProps = ITruncatedFormatProps;
-/** @deprecated use TruncatedFormatProps */
-export interface ITruncatedFormatProps extends Props {
-    children?: string;
-
-    /**
-     * Should the component keep track of the truncation state of the string content. If true, the
-     * value of `truncateLength` is ignored. When combined with a `showPopover` value of
-     * `WHEN_TRUNCATED`, popovers will only render when necessary.
-     *
-     * @default false;
-     */
-    detectTruncation?: boolean;
-
-    /**
-     * Values to use for character width, line height, cell padding, and buffer lines desired, when using approximate truncation.
-     * These values are used to guess at the size of the text and determine if the popover should be drawn. They should work well
-     * enough for default table styles, but may need to be overridden for more accuracy if the default styles or font size, etc
-     * are changed.
-     */
-    measureByApproxOptions?: ITrucatedFormateMeasureByApproximateOptions;
-
-    /**
-     * Height of the parent cell. Used by shouldComponentUpdate only
-     */
-    parentCellHeight?: number;
-
-    /**
-     * Width of the parent cell. Used by shouldComponentUpdate only
-     */
-    parentCellWidth?: number;
-
-    /**
-     * Sets the popover content style to `white-space: pre` if `true` or
-     * `white-space: normal` if `false`.
-     *
-     * @default false
-     */
-    preformatted?: boolean;
-
-    /**
-     * Configures when the popover is shown with the `TruncatedPopoverMode` enum.
-     *
-     * The enum values are:
-     * - `ALWAYS`: show the popover.
-     * - `NEVER`: don't show the popover.
-     * - `WHEN_TRUNCATED`: show the popover only when the text is truncated (default).
-     * - `WHEN_TRUNCATED_APPROX`: show the popover only when the text is trunctated, but use
-     *   a formula to calculate this based on text length, which is faster but less accurate.
-     *
-     * @default WHEN_TRUNCATED
-     */
-    showPopover?: TruncatedPopoverMode;
-
-    /**
-     * Number of characters that are displayed before being truncated and appended with the
-     * `truncationSuffix` prop. A value of 0 will disable truncation. This prop is ignored if
-     * `detectTruncation` is `true`.
-     *
-     * @default 2000
-     */
-    truncateLength?: number;
-
-    /**
-     * The string that is appended to the display string if it is truncated.
-     *
-     * @default "..."
-     */
-    truncationSuffix?: string;
-}
-
-export interface ITruncatedFormatState {
-    isTruncated?: boolean;
-    isPopoverOpen?: boolean;
-}
-
-/** @deprecated use TruncatedFormat2 */
-export class TruncatedFormat extends React.PureComponent<TruncatedFormatProps, ITruncatedFormatState> {
-    public static displayName = `${DISPLAYNAME_PREFIX}.TruncatedFormat`;
+export class TruncatedFormat2 extends React.PureComponent<TruncatedFormatProps, ITruncatedFormatState> {
+    public static displayName = `${DISPLAYNAME_PREFIX}.TruncatedFormat2`;
 
     public static defaultProps: TruncatedFormatProps = {
         detectTruncation: false,
@@ -225,7 +97,7 @@ export class TruncatedFormat extends React.PureComponent<TruncatedFormatProps, I
     private renderPopover() {
         const { children, preformatted } = this.props;
 
-        // `<Popover>` will always check the content's position on update
+        // `<Popover2>` will always check the content's position on update
         // regardless if it is open or not. This negatively affects perf due to
         // layout thrashing. So instead we manage the popover state ourselves
         // and mimic its popover target
@@ -236,19 +108,19 @@ export class TruncatedFormat extends React.PureComponent<TruncatedFormatProps, I
             );
             const popoverContent = <div className={popoverClasses}>{children}</div>;
             return (
-                <Popover
+                <Popover2
                     className={Classes.TABLE_TRUNCATED_POPOVER_TARGET}
-                    modifiers={{ preventOverflow: { boundariesElement: "window" } }}
                     content={popoverContent}
-                    position={Position.BOTTOM}
                     isOpen={true}
                     onClose={this.handlePopoverClose}
+                    placement="bottom"
+                    rootBoundary="document"
                 >
                     <Icon icon="more" />
-                </Popover>
+                </Popover2>
             );
         } else {
-            // NOTE: This structure matches what `<Popover>` does internally. If `<Popover>` changes, this must be updated.
+            // NOTE: This structure matches what `<Popover2>` does internally. If `<Popover2>` changes, this must be updated.
             return (
                 <span className={Classes.TABLE_TRUNCATED_POPOVER_TARGET} onClick={this.handlePopoverOpen}>
                     <Icon icon="more" />
