+++ packages/table/src/common/context.ts
@@ -16,10 +16,10 @@
 import * as PropTypes from "prop-types";
 import * as React from "react";
 
-export interface IColumnInteractionBarContextTypes {
+export interface ColumnInteractionBarContextTypes {
     enableColumnInteractionBar: boolean | null | undefined;
 }
 
-export const columnInteractionBarContextTypes: React.ValidationMap<IColumnInteractionBarContextTypes> = {
+export const columnInteractionBarContextTypes: React.ValidationMap<ColumnInteractionBarContextTypes> = {
     enableColumnInteractionBar: PropTypes.bool,
 };
