+++ packages/table/test/index.ts
@@ -20,12 +20,15 @@ import "./batcherTests.tsx";
 import "./cellTests.tsx";
 import "./clipboardTests.ts";
 import "./columnHeaderCellTests.tsx";
+import "./columnHeaderCell2Tests.tsx";
 import "./columnTests.tsx";
 import "./common/internal/";
 import "./editableCellTests.tsx";
 import "./editableCell2Tests.tsx";
 import "./editableNameTests.tsx";
-import "./formatsTests.tsx";
+import "./formats/truncatedFormatTests.tsx";
+import "./formats/truncatedFormat2Tests.tsx";
+import "./formats/jsonFormatTests.tsx";
 import "./gridTests.ts";
 import "./guidesTests.tsx";
 import "./loadableContentTests.tsx";
